+++ react-js-playground/src/components/Cockpit/Cockpit.css
@@ -1,0 +1,31 @@
+.red {
+	color: indianred;
+}
+
+.bold {
+	font-weight: bold;
+}
+
+.Cockpit button {
+	background-color: mediumseagreen;
+	color: white;
+	font: inherit;
+	border: 2px solid black;
+	padding: 8px;
+	cursor: pointer;
+	box-shadow: 0 2px 5px darkGray;
+	border-radius: 10px;
+}
+
+.Cockpit button:hover {
+	background-color: lightgreen;
+	color: black;
+}
+
+.Cockpit button.Red {
+	background-color: indianred;
+}
+
+.Cockpit button.Red:hover {
+	background-color: lightcoral;
+}
