+++ react-js-playground/src/components/Cockpit/Cockpit.js
@@ -1,10 +1,30 @@
 import React from 'react'
+import classes from './Cockpit.css'
 
-const Cockpit = (props) => {
+const cockpit = (props) => {
+	const assignedClasses = []
+
+	let btnClass = ''
+
+	if (props.showPersons) {
+		btnClass = classes.Red
+	}
+
+	if (props.persons.length <= 2) {
+		assignedClasses.push(classes.red) // classes = ['red']
+	}
+	if (props.persons.length <= 1) {
+		assignedClasses.push(classes.bold) // classes = ['red', 'bold']
+	}
+
 	return (
-		<div>
-			<p></p>
+		<div className={classes.Cockpit}>
+			<h1>Hi, I'm a React App</h1>
+			<p className={assignedClasses.join(' ')}>This is really working!</p>
+			<button className={btnClass} onClick={props.togglePersons}>
+				{props.buttonText}
+			</button>
 		</div>
 	)
 }
-export default Cockpit
+export default cockpit
