+++ react-js-playground/src/containers/App.css
@@ -2,38 +2,6 @@
 	text-align: center;
 }
 
-.red {
-	color: indianred;
-}
-
-.bold {
-	font-weight: bold;
-}
-
-.App button {
-	background-color: mediumseagreen;
-	color: white;
-	font: inherit;
-	border: 2px solid black;
-	padding: 8px;
-	cursor: pointer;
-	box-shadow: 0 2px 5px darkGray;
-	border-radius: 10px;
-}
-
-.App button:hover {
-	background-color: lightgreen;
-	color: black;
-}
-
-.App button.Red {
-	background-color: indianred;
-}
-
-.App button.Red:hover {
-	background-color: lightcoral;
-}
-
 /* Use the global selector to specify classes to be part of global scope 
 Example:
 :global .App {...}
