+++ react-js-playground/src/containers/App.js
@@ -57,47 +57,36 @@ class App extends Component {
 
 	render() {
 		//! Anything that is rendered to the screen including logic
-		//! that affects rendered components goes in the render()
+		//! that affects rendered components/state goes in the render()
 		//! method body
 		let buttonText = 'Show Names'
 
 		let persons = null
 
-		let btnClass = ''
-
 		// Code to manage the 'Persons' array (list)
 		if (this.state.showPersons) {
 			persons = (
-				<div>
-					<Persons
-						persons={this.state.persons}
-						clicked={this.deletePersonHandler}
-						changed={this.nameChangedHandler}
-					/>
-				</div>
+				<Persons
+					persons={this.state.persons}
+					clicked={this.deletePersonHandler}
+					changed={this.nameChangedHandler}
+				/>
 			)
+
 			// Dynamically styling output if 'showPersons' is true
 			buttonText = 'Hide Names'
-			btnClass = classes.Red
-		}
-
-		const assignedClasses = []
-		if (this.state.persons.length <= 2) {
-			assignedClasses.push(classes.red) // classes = ['red']
-		}
-		if (this.state.persons.length <= 1) {
-			assignedClasses.push(classes.bold) // classes = ['red', 'bold']
 		}
 
 		//! JSX/React components that are rendered to the screen go
 		//! in the return statement within the render() method
 		return (
 			<div className={classes.App}>
-				<h1>Hi, I'm a React App</h1>
-				<p className={assignedClasses.join(' ')}>This is really working!</p>
-				<button className={btnClass} onClick={this.togglePersonsHandler}>
-					{buttonText}
-				</button>
+				<Cockpit
+					persons={this.state.persons}
+					showPersons={this.state.showPersons}
+					togglePersons={this.togglePersonsHandler}
+					buttonText={buttonText}
+				/>
 				{persons}
 			</div>
 		)
